move test outside scr folder
